+++ R/cv.hfr.R
@@ -16,17 +16,19 @@
 #'
 #' @param x Input matrix or data.frame, of dimension \eqn{(N\times p)}{(N x p)}; each row is an observation vector.
 #' @param y Response variable.
+#' @param weights an optional vector of weights to be used in the fitting process. Should be NULL or a numeric vector. If non-NULL, weighted least squares is used for the level-specific regressions.
 #' @param kappa_grid A vector of target effective degrees of freedom of the regression.
 #' @param q Thinning parameter representing the quantile cut-off (in terms of contributed variance) above which to consider levels in the hierarchy. This can used to reduce the number of levels in high-dimensional problems. Default is no thinning.
 #' @param intercept Should intercept be fitted. Default is \code{intercept=TRUE}.
 #' @param standardize Logical flag for \code{x} variable standardization prior to fitting the model. The coefficients are always returned on the original scale. Default is \code{standardize=TRUE}.
 #' @param nfolds The number of folds for k-fold cross validation. Default is \code{nfolds=10}.
 #' @param foldid An optional vector of values between \code{1} and \code{nfolds} identifying what fold each observation is in. If supplied, \code{nfolds} can be missing.
+#' @param partial_method Indicate whether to use pairwise partial correlations, or shrinkage partial correlations.
 #' @param ...  Additional arguments passed to \code{hclust}.
 #' @return A 'cv.hfr' regression object.
 #' @author Johann Pfitzinger
 #' @references
-#' Pfitzinger, J. (2021).
+#' Pfitzinger, J. (2022).
 #' Cluster Regularization via a Hierarchical Feature Regression.
 #' arXiv 2107.04831[statML]
 #'
@@ -38,7 +40,7 @@
 #'
 #' @export
 #'
-#' @seealso \code{hfr}, \code{coef} and \code{predict} methods
+#' @seealso \code{\link{hfr}}, \code{\link{coef}}, \code{\link{plot}} and \code{\link{predict}} methods
 #'
 #' @importFrom quadprog solve.QP
 #' @importFrom stats sd
@@ -47,12 +49,14 @@
 cv.hfr <- function(
   x,
   y,
+  weights = NULL,
   kappa_grid = seq(0, 1, by = 0.1),
   q = NULL,
   intercept = TRUE,
   standardize = TRUE,
   nfolds = 10,
   foldid = NULL,
+  partial_method = c("pairwise", "shrinkage"),
   ...
 ) {
 
@@ -81,6 +85,20 @@ cv.hfr <- function(
     stop("each 'kappa' must be between 0 and 1.")
   }
 
+  if (!is.null(weights)) {
+    if (length(weights) != nobs)
+      stop("'weights' must have same length as 'y'")
+    if (any(is.na(weights)))
+      stop("'NA' values in 'weights'")
+    if (any(weights < 0))
+      stop("'weights' can only contain positive numerical values")
+    wts <- sqrt(weights)
+  } else {
+    wts <- rep(1, nobs)
+  }
+
+  partial_method = match.arg(partial_method)
+
   if (is.null(foldid))
     foldid = sample(rep(seq(nfolds), length = nobs))
   else nfolds = max(foldid)
@@ -88,7 +106,7 @@ cv.hfr <- function(
   # Get feature names
   var_names <- colnames(x)
   if (is.null(var_names)) var_names <- paste("X", 1:ncol(x), sep = ".")
-  if (intercept) var_names <- c("intercept", var_names)
+  if (intercept) var_names <- c("(Intercept)", var_names)
 
   # Convert 'x' to matrix
   x <- data.matrix(x)
@@ -122,7 +140,7 @@ cv.hfr <- function(
         xs = x_fit
       }
 
-      v = .get_level_reg(xs, y_fit, nvars, nobs, q, intercept, ...)
+      v = .get_level_reg(xs, y_fit, wts[!ix], nvars, nobs, q, intercept, partial_method, ...)
       meta_opt <- .get_meta_opt(y_fit, kappa_grid, nvars, nobs, var_names, standardize, intercept, standard_sd, standard_mean, v)
 
       beta_mat <- meta_opt$beta
@@ -154,7 +172,7 @@ cv.hfr <- function(
     xs <- x
   }
 
-  v = .get_level_reg(xs, y, nvars, nobs, q, intercept, ...)
+  v = .get_level_reg(xs, y, wts, nvars, nobs, q, intercept, partial_method, ...)
   meta_opt <- .get_meta_opt(y, kappa_grid, nvars, nobs, var_names, standardize, intercept, standard_sd, standard_mean, v)
 
   beta_mat <- meta_opt$beta