+++ NEWS.md
@@ -1,3 +1,9 @@
 ## hfr 0.5.0
 
 - Note that this starts from version `hfr 0.5.0`.
+
+## hfr 0.6.0
+
+- The plotting functionality has been expanded to highlight statistically insignificant branches
+- Option to compute partial correlations using the pairwise (default) approach, or shrinkage approach
+- Optional 'weights' argument to implement a weighted version of HFR.